Clarified question for week 51
Originally the example didn't make sense because it asked for the element that occurs greater than half of the time. In the example '1' only occurs three times thus not making it the majority element by the rules. If they meant a majority of the time up to that point in the list, then you wouldn't use the length of the entire list to calc the majority as the example showed. Thus I changed it to greater than or equal to floor(half).